+++ Tests/GRDBTests/MutablePersistableRecordChangesTests.swift
@@ -561,4 +561,64 @@ class MutablePersistableRecordChangesTests: GRDBTestCase {
             }
         }
     }
+    
+    func testDatabaseChangesWithClass() throws {
+        class Player: Encodable, EncodableRecord {
+            var id: Int64
+            var name: String
+            var score: Int
+            
+            init(id: Int64, name: String, score: Int) {
+                self.id = id
+                self.name = name
+                self.score = score
+            }
+        }
+        
+        var player = Player(id: 1, name: "Arthur", score: 1000)
+        do {
+            let changes = try player.databaseChanges { _ in }
+            XCTAssert(changes.isEmpty)
+        }
+        do {
+            let changes = try player.databaseChanges {
+                $0.name = "Barbara"
+            }
+            XCTAssertEqual(changes, ["name": "Arthur".databaseValue])
+        }
+        do {
+            let changes = try player.databaseChanges {
+                $0.name = "Craig"
+                $0.score = 200
+            }
+            XCTAssertEqual(changes, ["name": "Barbara".databaseValue, "score": 1000.databaseValue])
+        }
+    }
+    
+    func testDatabaseChangesWithStruct() throws {
+        struct Player: Encodable, EncodableRecord {
+            var id: Int64
+            var name: String
+            var score: Int
+        }
+        
+        var player = Player(id: 1, name: "Arthur", score: 1000)
+        do {
+            let changes = try player.databaseChanges { _ in }
+            XCTAssert(changes.isEmpty)
+        }
+        do {
+            let changes = try player.databaseChanges {
+                $0.name = "Barbara"
+            }
+            XCTAssertEqual(changes, ["name": "Arthur".databaseValue])
+        }
+        do {
+            let changes = try player.databaseChanges {
+                $0.name = "Craig"
+                $0.score = 200
+            }
+            XCTAssertEqual(changes, ["name": "Barbara".databaseValue, "score": 1000.databaseValue])
+        }
+    }
 }